Modernize ESI.NET: netstandard2.0/net8.0/net10.0, ESI compatibility-date versioning, full endpoint coverage

+++ ESI.NET/Models/Alliance/Alliance.cs
@@ -6,19 +6,19 @@ namespace ESI.NET.Models.Alliance
     public class Alliance
     {
         [JsonProperty("creator_corporation_id")]
-        public int CreatorCorporationId { get; set; }
+        public long CreatorCorporationId { get; set; }
 
         [JsonProperty("creator_id")]
-        public int CreatorId { get; set; }
+        public long CreatorId { get; set; }
 
         [JsonProperty("date_founded")]
         public DateTime DateFounded { get; set; }
 
         [JsonProperty("executor_corporation_id")]
-        public int ExecutorCorporationId { get; set; }
+        public long ExecutorCorporationId { get; set; }
 
         [JsonProperty("faction_id")]
-        public int FactionId { get; set; }
+        public long FactionId { get; set; }
 
         [JsonProperty("name")]
         public string Name { get; set; }
