+++ ESI.NET/Models/Character/Affiliation.cs
@@ -5,15 +5,15 @@ namespace ESI.NET.Models.Character
     public class Affiliation
     {
         [JsonProperty("alliance_id")]
-        public int AllianceId { get; set; }
+        public long AllianceId { get; set; }
 
         [JsonProperty("character_id")]
-        public int CharacterId { get; set; }
+        public long CharacterId { get; set; }
 
         [JsonProperty("corporation_id")]
-        public int CorporationId { get; set; }
+        public long CorporationId { get; set; }
 
         [JsonProperty("faction_id")]
-        public int FactionId { get; set; }
+        public long FactionId { get; set; }
     }
 }
