+++ ESI.NET/Models/Corporation/AllianceHistory.cs
@@ -6,13 +6,13 @@ namespace ESI.NET.Models.Corporation
     public class AllianceHistory
     {
         [JsonProperty("alliance_id")]
-        public int AllianceId { get; set; }
+        public long AllianceId { get; set; }
 
         [JsonProperty("is_deleted")]
         public bool IsDeleted { get; set; }
 
         [JsonProperty("record_id")]
-        public int RecordId { get; set; }
+        public long RecordId { get; set; }
 
         [JsonProperty("start_date")]
         public DateTime StartDate { get; set; }
