+++ ESI.NET/Models/Corporation/IssuedMedal.cs
@@ -6,16 +6,16 @@ namespace ESI.NET.Models.Corporation
     public class IssuedMedal
     {
         [JsonProperty("character_id")]
-        public int CharacterId { get; set; }
+        public long CharacterId { get; set; }
 
         [JsonProperty("issued_at")]
         public DateTime IssuedAt { get; set; }
 
         [JsonProperty("issuer_id")]
-        public int IssuerId { get; set; }
+        public long IssuerId { get; set; }
 
         [JsonProperty("medal_id")]
-        public int MedalId { get; set; }
+        public long MedalId { get; set; }
 
         [JsonProperty("reason")]
         public string Reason { get; set; }
