+++ ESI.NET/Models/Killmails/Information.cs
@@ -16,34 +16,34 @@ namespace ESI.NET.Models.Killmails
         public DateTime KillmailTime { get; set; }
 
         [JsonProperty("moon_id")]
-        public int MoonId { get; set; }
+        public long MoonId { get; set; }
 
         [JsonProperty("solar_system_id")]
-        public int SolarSystemId { get; set; }
+        public long SolarSystemId { get; set; }
 
         [JsonProperty("victim")]
         public Victim Victim { get; set; }
 
         [JsonProperty("war_id")]
-        public int WarId { get; set; }
+        public long WarId { get; set; }
     }
 
     public class Attacker
     {
         [JsonProperty("alliance_id")]
-        public int AllianceId { get; set; }
+        public long AllianceId { get; set; }
 
         [JsonProperty("character_id")]
-        public int CharacterId { get; set; }
+        public long CharacterId { get; set; }
 
         [JsonProperty("corporation_id")]
-        public int CorporationId { get; set; }
+        public long CorporationId { get; set; }
 
         [JsonProperty("damage_done")]
-        public int DamageDone { get; set; }
+        public long DamageDone { get; set; }
 
         [JsonProperty("faction_id")]
-        public int FactionId { get; set; }
+        public long FactionId { get; set; }
 
         [JsonProperty("final_blow")]
         public bool FinalBlow { get; set; }
@@ -52,31 +52,31 @@ namespace ESI.NET.Models.Killmails
         public double SecurityStatus { get; set; }
 
         [JsonProperty("ship_type_id")]
-        public int ShipTypeId { get; set; }
+        public long ShipTypeId { get; set; }
 
         [JsonProperty("weapon_type_id")]
-        public int WeaponTypeId { get; set; }
+        public long WeaponTypeId { get; set; }
     }
 
     public class Victim
     {
         [JsonProperty("alliance_id")]
-        public int AllianceId { get; set; }
+        public long AllianceId { get; set; }
 
         [JsonProperty("character_id")]
-        public int CharacterId { get; set; }
+        public long CharacterId { get; set; }
 
         [JsonProperty("corporation_id")]
-        public int CorporationId { get; set; }
+        public long CorporationId { get; set; }
 
         [JsonProperty("damage_taken")]
-        public int DamageTaken { get; set; }
+        public long DamageTaken { get; set; }
 
         [JsonProperty("faction_id")]
-        public int FactionId { get; set; }
+        public long FactionId { get; set; }
 
         [JsonProperty("ship_type_id")]
-        public int ShipTypeId { get; set; }
+        public long ShipTypeId { get; set; }
 
         [JsonProperty("items")]
         public List<Item> Items { get; set; } = new List<Item>();
@@ -88,10 +88,10 @@ namespace ESI.NET.Models.Killmails
     public class Item
     {
         [JsonProperty("flag")]
-        public int Flag { get; set; }
+        public long Flag { get; set; }
 
         [JsonProperty("item_type_id")]
-        public int ItemTypeId { get; set; }
+        public long ItemTypeId { get; set; }
 
         [JsonProperty("items")]
         public List<Item> Items { get; set; }
@@ -103,6 +103,6 @@ namespace ESI.NET.Models.Killmails
         public long QuantityDropped { get; set; }
 
         [JsonProperty("singleton")]
-        public int Singleton { get; set; }
+        public long Singleton { get; set; }
     }
 }
