+++ ESI.NET/Models/Location/Ship.cs
@@ -5,7 +5,7 @@ namespace ESI.NET.Models.Location
     public class Ship
     {
         [JsonProperty("ship_type_id")]
-        public int ShipTypeId { get; set; }
+        public long ShipTypeId { get; set; }
 
         [JsonProperty("ship_item_id")]
         public long ShipItemId { get; set; }
