+++ ESI.NET/Models/Mail/MailingList.cs
@@ -5,7 +5,7 @@ namespace ESI.NET.Models.Mail
     public class MailingList
     {
         [JsonProperty("mailing_list_id")]
-        public int MailingListId { get; set; }
+        public long MailingListId { get; set; }
 
         [JsonProperty("name")]
         public string Name { get; set; }
