+++ ESI.NET/Models/Mail/Message.cs
@@ -9,7 +9,7 @@ namespace ESI.NET.Models.Mail
         public string Subject { get; set; }
 
         [JsonProperty("from")]
-        public int From { get; set; }
+        public long From { get; set; }
 
         [JsonProperty("timestamp")]
         public string Timestamp { get; set; }
@@ -33,6 +33,6 @@ namespace ESI.NET.Models.Mail
         public string RecipientType { get; set; }
 
         [JsonProperty("recipient_id")]
-        public int RecipientId { get; set; }
+        public long RecipientId { get; set; }
     }
 }
