+++ ESI.NET/Models/PlanetaryInteraction/Schematic.cs
@@ -5,7 +5,7 @@ namespace ESI.NET.Models.PlanetaryInteraction
     public class Schematic
     {
         [JsonProperty("cycle_time")]
-        public int CycleTime { get; set; }
+        public long CycleTime { get; set; }
 
         [JsonProperty("schematic_name")]
         public string Name { get; set; }
