+++ ESI.NET/Models/Status/Status.cs
@@ -6,7 +6,7 @@ namespace ESI.NET.Models.Status
     public class Status
     {
         [JsonProperty("players")]
-        public int Players { get; set; }
+        public long Players { get; set; }
 
         [JsonProperty("server_version")]
         public string ServerVersion { get; set; }
