+++ ESI.NET/Models/Wallet/Wallet.cs
@@ -5,7 +5,7 @@ namespace ESI.NET.Models.Wallet
     public class Wallet
     {
         [JsonProperty("division")]
-        public int Division { get; set; }
+        public long Division { get; set; }
 
         [JsonProperty("balance")]
         public decimal Balance { get; set; }
