+++ ESI.NET/Models/_SSO/AuthorizedCharacterData.cs
@@ -4,7 +4,7 @@ namespace ESI.NET.Models.SSO
 {
     public class AuthorizedCharacterData
     {
-        public int CharacterID { get; set; }
+        public long CharacterID { get; set; }
         public string CharacterName { get; set; }
         public DateTime ExpiresOn { get; set; }
         public string Scopes { get; set; }
@@ -13,8 +13,8 @@ namespace ESI.NET.Models.SSO
 
         public string Token { get; set; }
         public string RefreshToken { get; set; }
-        public int AllianceID { get; set; }
-        public int CorporationID { get; set; }
-        public int FactionID { get; set; }
+        public long AllianceID { get; set; }
+        public long CorporationID { get; set; }
+        public long FactionID { get; set; }
     }
 }
